% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascii.R
\name{chr_to_ascii}
\alias{chr_to_ascii}
\title{Convert a \code{character} Vector to ASCII Encoding}
\usage{
chr_to_ascii(chr, trans = TRUE)
}
\arguments{
\item{trans}{Should non-ASCII characters be transliterated?}

\item{string}{A \code{character} vector}
}
\value{
The input, converted to ASCII representation
}
\description{
Convert a \code{character} Vector to ASCII Encoding
}
\keyword{internal}
