% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_redcap_records.R
\name{download-case-projects}
\alias{download-case-projects}
\alias{download_nit}
\alias{download_nca}
\title{Download New Interview Tool Data}
\usage{
download_nit(
  api_token = Sys.getenv("redcap_NIT_token"),
  headers = c("raw", "label"),
  values = c("label", "raw"),
  filter = NULL,
  dir = "V:/EPI DATA ANALYTICS TEAM/COVID SANDBOX REDCAP DATA/Data for R/nit/",
  file = paste0("nit_data_", str_date(), ".csv"),
  force = FALSE
)

download_nca(
  api_token = Sys.getenv("redcap_NCA_token"),
  headers = c("raw", "label"),
  values = c("label", "raw"),
  filter = NULL,
  dir = "V:/EPI DATA ANALYTICS TEAM/COVID SANDBOX REDCAP DATA/Data for R/nca/",
  file = paste0("nca_data_", str_date(), ".csv"),
  force = FALSE
)
}
\arguments{
\item{api_token}{\code{character}. API token/key for accessing the project
programmatically. Best practice is to set as an environment variable and
retrieve using \code{Sys.getenv("token_environment_variable")}. See
\code{\link{env-variables}} for more information.}

\item{headers}{\code{"raw"} or \code{"label"}. Should column names ("headers") be
exported as the raw variable names (\code{"raw"}, the default) or the label
shown in the UI (\code{"label"})?}

\item{values}{\code{"raw"} or \code{"label"}. Should values in multiple-choice and
checkbox fields be exported as the raw coded value (\code{"raw"}, the default)
or the label for the coded value (\code{"label"})?}

\item{filter}{\code{character}. REDcap filtering logic to apply prior to download.
This must be in the REDcap logic syntax.}

\item{dir}{\code{character}. Directory to save data; this should usually remain
unchanged.}

\item{file}{\code{character}. File name for data; this should usually remain
unchanged. The resulting file will \strong{always} have a "csv" extension; if
you pass a \code{file} name without a "csv" extension, the function will
replace it with a warning.}

\item{force}{Should the download overwrite an existing file, if one exists?
The default is \code{FALSE}, which errors if an existing file is found.}
}
\value{
The path to the data as an \code{fs_path} character vector
}
\description{
\code{download_nit()} downloads records from the New Interview Tool (NIT) project

\code{download_nca()} downloads records from the New Case Assignment (NCA) project
}
\examples{
\dontrun{
  # Download the current NIT data
  download_nit()

  # Download NIT and load immediately
  nit_data <- load_nit(download_nit())

  # Do the same with pipe (`\%>\%`)
  nit_data <- download_nit() \%>\% load_nit()

  # Overwrite the previous file
  download_nit(force = TRUE)

  # Download NCA data
  download_nca()

  # Download NCA and overwrite previous file
  nca_data <- download_nca() \%>\% load_nca()
}
}
