% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_redcap_records.R
\name{download_redcap_records}
\alias{download_redcap_records}
\title{Download New Interview Tool Data}
\usage{
download_redcap_records(
  api_token,
  dir,
  file,
  headers = c("raw", "label"),
  values = c("raw", "label"),
  filter = NULL,
  force = FALSE
)
}
\arguments{
\item{api_token}{\code{character}. API token/key for accessing the project
programmatically. Best practice is to set as an environment variable and
retrieve using \code{Sys.getenv("token_environment_variable")}. See
\code{\link{env-variables}} for more information.}

\item{dir}{\code{character}. Directory to save data; this should usually remain
unchanged.}

\item{file}{\code{character}. File name for data; this should usually remain
unchanged. The resulting file will \strong{always} have a "csv" extension; if
you pass a \code{file} name without a "csv" extension, the function will
replace it with a warning.}

\item{headers}{\code{"raw"} or \code{"label"}. Should column names ("headers") be
exported as the raw variable names (\code{"raw"}, the default) or the label
shown in the UI (\code{"label"})?}

\item{values}{\code{"raw"} or \code{"label"}. Should values in multiple-choice and
checkbox fields be exported as the raw coded value (\code{"raw"}, the default)
or the label for the coded value (\code{"label"})?}

\item{filter}{\code{character}. REDcap filtering logic to apply prior to download.
This must be in the REDcap logic syntax.}

\item{force}{Should the download overwrite an existing file, if one exists?
The default is \code{FALSE}, which errors if an existing file is found.}
}
\value{
The path to the data as an \code{fs_path} character vector
}
\description{
\code{download_redcap_records()} downloads records from a REDcap project
}
