% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_redcap_records.R
\name{load-case-projects}
\alias{load-case-projects}
\alias{load_nit}
\alias{load_nca}
\title{Load Last Downloaded Data for a REDcap Project}
\usage{
load_nit(date = NULL, path = path_nit(date), clean_names = TRUE, ...)

load_nca(date = NULL, path = path_nca(date), clean_names = TRUE, ...)
}
\arguments{
\item{date}{The download date of the file to read; the default (\code{NULL}) will
read the most recent file}

\item{path}{The path to the file to read; if supplied, the function will read
this file instead}

\item{clean_names}{Should column names be cleaned by
\code{\link[janitor:clean_names]{clean_names()}}?}

\item{...}{Additional parameters to pass to
\code{\link[covidprod:read_file_delim]{read_file_delim()}}}
}
\value{
A \code{tibble} if reading one file; a list of \code{tibble}s if reading
multiple
}
\description{
\code{load_nit()} and \code{load_nca()} load the latest data downloaded from the
respective REDcap project. To find the file paths to these data, use
\code{\link[covidprod:path_nit]{path_nit()}} or
\code{\link[covidprod:path_nit]{path_nca()}}. To download the data, use
\code{\link[covidprod:download_nit]{download_nit()}} or
\code{\link[covidprod:download_nca]{download_nca()}}.
}
