% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-date_tbl.R
\name{num_dt_dttm}
\alias{num_dt_dttm}
\title{Get Best Type for a Date Numeric}
\usage{
num_dt_dttm(date)
}
\arguments{
\item{date}{A date in numeric format}
}
\value{
\code{"date"}, \code{"datetime"}, or \code{NA_character_}
}
\description{
\code{num_dt_dttm()} returns the type that results in a datetime closer to the
current datetime. It returns \code{NA_character_} if the input is a missing value.
}
