% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_by_date.R
\name{path_by_date}
\alias{path_by_date}
\title{Return the Path to a NBS Investigations File}
\usage{
path_by_date(
  dir,
  date_format,
  date_regex,
  date = NULL,
  file_regex = ".*{date}.*",
  type = "any",
  force_latest = TRUE
)
}
\arguments{
\item{dir}{The directory holding the files of interest}

\item{date_format}{The format of the date in the file name; see
\code{\link[base:format]{format()}} for more information}

\item{date_regex}{A regular expression matching the date in the file name.
This will hopefully be deprecated in the future, but is currently needed
for extracting dates from the file path strings.}

\item{date}{A \code{Date} or string in the format "YYYY-MM-DD"}

\item{file_regex}{A regular expression matching the file names of interest.
The location of the date in the file name is specified using \code{"{date}"}, as
in the glue package.}

\item{type}{The file type(s) to return, one or more of \code{"any"}, \code{"file"},
\code{"directory"}, \code{"symlink"}, \code{"FIFO"}, \code{"socket"}, \code{"character_device"} or
\code{"block_device"}.}

\item{force_latest}{If multiple files with the given \code{date} are found,
should the function return only the file with the latest creation date
(\code{TRUE}, the default), or should it return all file paths (\code{FALSE})?}
}
\value{
An \code{fs_path} character vector
}
\description{
\verb{path_*} functions return paths to one or more files of interest
containing the supplied date in their names and ending with the supplied
file extension. These are convenience wrappers around \code{path_by_date()}, which
returns path that contain a specified date given additional parameters.
}
\details{
\itemize{
\item{\code{path_nit()} returns path(s) to NIT data}
\item{\code{path_nca()} returns path(s) to NCA data}
}
}
