% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_create.R
\name{path_clean}
\alias{path_clean}
\title{Clean a Path with the fs Package}
\usage{
path_clean(path, home = c("r", "user"))
}
\arguments{
\item{path}{A character vector of one or more paths.}

\item{home}{The home directory definition to use when expanding a path. The
default is to use \strong{R}'s definition, but the system definition of the
user's home directory can also be used. These are equivalent to calling
\code{path_expand_r()} or \code{path_expand()}, respectively.}
}
\value{
The cleaned path(s) in an \code{fs_path} object, which is a character
vector that also has class \code{fs_path}
}
\description{
\code{path_clean()} is wrapper around \code{\link[fs:path_tidy]{path_tidy()}},
\code{\link[fs:path_expand]{path_expand_r()}}
(or \code{\link[fs:path_expand]{path_expand()}}), and
\code{\link[fs:path_math]{path_norm()}}. It tidies, then expands, then
normalizes the given path(s).
}
