% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_create.R
\name{path_create}
\alias{path_create}
\title{Create a Path from Components}
\usage{
path_create(..., ext = NULL, home = c("r", "user"))
}
\arguments{
\item{...}{character vectors, if any values are NA, the result will also be
NA. The paths follow the recycling rules used in the tibble package,
namely that only length 1 arguments are recycled.}

\item{ext}{An optional extension to append to the generated path.}

\item{home}{The home directory definition to use when expanding a path. The
default is to use \strong{R}'s definition, but the system definition of the
user's home directory can also be used. These are equivalent to calling
\code{path_expand_r()} or \code{path_expand()}, respectively.}
}
\value{
The combined and cleaned path in an \code{fs_path} object, which is a
character vector that also has class \code{fs_path}
}
\description{
\code{path_create()} creates a cleaned path from the components specified in \code{...}
and an optional file extension (\code{ext}). The path is automatically expanded
using either the \strong{R} or system definition of the home directory.
}
