% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file_delim.R
\name{read_file_delim}
\alias{read_file_delim}
\title{Efficently Read Delimited Files}
\usage{
read_file_delim(
  file,
  col_select = vroom::everything(),
  col_types = vroom::cols(.default = vroom::col_character()),
  na = c("", ".", "NA", "na", "Na", "N/A", "n/a", "N/a", "NULL", "null", "Null"),
  guess_max = .Machine$integer.max\%/\%100L,
  delim = NULL,
  ...
)
}
\arguments{
\item{file}{path to a local file.}

\item{col_select}{One or more selection expressions, like in
\code{dplyr::select()}. Use \code{c()} or \code{list()} to use more than one expression.
See \code{?dplyr::select} for details on available selection options.}

\item{col_types}{One of \code{NULL}, a \code{\link[readr:cols]{cols()}} specification, or
a string. See \code{vignette("readr")} for more details.

If \code{NULL}, all column types will be imputed from the first 1000 rows
on the input. This is convenient (and fast), but not robust. If the
imputation fails, you'll need to supply the correct types yourself.

If a column specification created by \code{\link[readr:cols]{cols()}}, it must contain
one column specification for each column. If you only want to read a
subset of the columns, use \code{\link[readr:cols]{cols_only()}}.

Alternatively, you can use a compact string representation where each
character represents one column:
c = character, i = integer, n = number, d = double,
l = logical, f = factor, D = date, T = date time, t = time, ? = guess, or
\code{_}/\code{-} to skip the column.}

\item{na}{Character vector of strings to interpret as missing values. Set this
option to \code{character()} to indicate no missing values.}

\item{guess_max}{Maximum number of records to use for guessing column types.}

\item{delim}{One or more characters used to delimit fields within a
file. If \code{NULL} the delimiter is guessed from the set of \code{c(",", "\t", " ", "|", ":", ";")}.}

\item{...}{Additional arguments to pass to \code{\link[vroom:vroom]{vroom()}}}
}
\value{
A \code{tibble} if reading one file; a list of \code{tibble}s if reading
multiple
}
\description{
\code{read_file_delim()} reads delimited files using
\code{\link[vroom:vroom]{vroom()}}. This allows the use of ALTREP columns,
which don't load data into memory until they are needed.
}
\details{
By default, \code{read_file_delim()} does not attempt to guess column types and
reads all columns as character. This can be changed by setting
\code{col_types = vroom::cols(.default = vroom::col_guess())}. If columns are
guessed, the default is to use all rows; this can be changed by setting
\code{guess_max} to a different value.

This saves a significant amount of time and space when loading data with many
rarely used columns.
}
