% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-date_tbl.R
\name{str_dt_dttm}
\alias{str_dt_dttm}
\title{Get Best Parsing Type for a Date String}
\usage{
str_dt_dttm(date)
}
\arguments{
\item{date}{A potential date encoded as a string}
}
\value{
\code{"date"}, \code{"datetime"}, or \code{NA_character_}
}
\description{
\code{str_dt_dttm()} returns \code{"date"} if the string can be parsed as a \code{Date} and
contains no time information. It returns \code{"datetime"} if the string can be
parsed as a datetime and contains time information. Otherwise, it returns
\code{NA_character_}.
}
\keyword{internal}
