% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-date_tbl.R
\name{str_is_dt}
\alias{str_is_dt}
\title{Check that a String Can Be Parsed as a Date (and not a datetime)}
\usage{
str_is_dt(date)
}
\arguments{
\item{date}{A string to check for date information}
}
\value{
A scalar \code{logical}
}
\description{
Check that a String Can Be Parsed as a Date (and not a datetime)
}
\keyword{internal}
