% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-date_tbl.R
\name{str_is_dttm}
\alias{str_is_dttm}
\title{Check That a String can be parsed as a datetime}
\usage{
str_is_dttm(date)
}
\arguments{
\item{date}{A string to check for date information}
}
\value{
A scalar \code{logical}
}
\description{
Check That a String can be parsed as a datetime
}
\keyword{internal}
