% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_prod.R
\name{summarize_prod}
\alias{summarize_prod}
\title{Summarize Productivity in NCA and NIT}
\usage{
summarize_prod(nca = load_nca(), nit = load_nit())
}
\arguments{
\item{nca}{NCA data, as output by
\code{\link[covidprod:load_nca]{load_nca()}}}

\item{nit}{NIT data, as output by
\code{\link[covidprod:load_nit]{load_nit()}}}
}
\value{
A \code{tibble} with one row per day since \code{"2020-02-02"}
}
\description{
Summarize Productivity in NCA and NIT
}
