% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-tidyeval-symbols.R
\name{utils-tidyeval-symbols}
\alias{utils-tidyeval-symbols}
\title{Symbols Used in Tidy Evaluation}
\description{
Import tidy eval pronouns (\code{\link[rlang:tidyeval-data]{.data}} and
\code{\link[rlang:tidyeval-data]{.env}}), as well as forcing operators
(\code{\link[rlang:nse-force]{!!}}, \code{\link[rlang:nse-force]{!!!}}, and
\code{\link[rlang:nse-force]{:=}}).
These operators cannot be called using the usual \code{::} syntax and must thus
be imported to prevent R CMD CHECK warnings.
}
\keyword{internal}
