% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-date_tbl.R
\name{validate_date_attr}
\alias{validate_date_attr}
\title{Validate than an Object has a \code{"date"} Attribute}
\usage{
validate_date_attr(x)
}
\arguments{
\item{x}{An object to check}
}
\value{
The input object if valid; otherwise, throws an error
}
\description{
Validate than an Object has a \code{"date"} Attribute
}
\keyword{internal}
