% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-date_tbl.R
\name{validate_date_range}
\alias{validate_date_range}
\title{Validate That a \code{date} falls in Valid Range}
\usage{
validate_date_range(date)
}
\arguments{
\item{date}{A \code{Date} or datetime}
}
\value{
The input if valid; otherwise, throws an error
}
\description{
\code{validate_date_range()} validates that a date falls between \code{"2020-01-01"}
and the current date (inclusive).
}
\keyword{internal}
