% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-date_tbl.R
\name{validate_date_type}
\alias{validate_date_type}
\title{Validate That a Potential \code{date} is a Valid Type}
\usage{
validate_date_type(date)
}
\arguments{
\item{date}{Potential \code{date} to check}
}
\value{
The input if valid; otherwise, throws an error
}
\description{
\code{validate_date_type()} checks that a potential \code{date} is either a \code{Date},
datetime, string, or numeric, and that it is a scalar.
}
\keyword{internal}
